Reduce the no of indexers from 4 to 3
