+++ src/module.indexer/model/_main.ts
@@ -11,8 +11,7 @@ import { JsonRpcClient } from '@defichain/jellyfish-api-jsonrpc'
 import { RawBlockMapper } from '@src/module.model/raw.block'
 import { NotFoundIndexerError } from '@src/module.indexer/error'
 import { blockchain as defid } from '@defichain/jellyfish-api-core'
-import { OracleAppointedIndexer } from '@src/module.indexer/model/oracle.appointed'
-import { OraclePriceFeedIndexer } from '@src/module.indexer/model/oracle.price.feed'
+import { OracleStatusIndexer } from '@src/module.indexer/model/oracle.status'
 import { OraclePriceDataIndexer } from '@src/module.indexer/model/oracle.price.data'
 import { OraclePriceAggregationIndexer } from '@src/module.indexer/model/oracle.price.aggregration'
 
@@ -33,8 +32,7 @@ export class MainIndexer {
     private readonly transaction: TransactionIndexer,
     private readonly transactionVin: TransactionVinIndexer,
     private readonly transactionVout: TransactionVoutIndexer,
-    private readonly oracleAppointed: OracleAppointedIndexer,
-    private readonly oraclePriceFeed: OraclePriceFeedIndexer,
+    private readonly oracleAppointed: OracleStatusIndexer,
     private readonly oraclePriceData: OraclePriceDataIndexer,
     private readonly oraclePriceAggregration: OraclePriceAggregationIndexer
   ) {
@@ -47,7 +45,6 @@ export class MainIndexer {
       transactionVin,
       transactionVout,
       oracleAppointed,
-      oraclePriceFeed,
       oraclePriceData,
       oraclePriceAggregration
     ]
